% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.plot.grid.R
\name{basic.plot.grid}
\alias{basic.plot.grid}
\title{Creates the shell of a plot showing relationship between jury pool preferences and jury verdict probabilities}
\usage{
basic.plot.grid(main, xlab, ylab)
}
\arguments{
\item{main}{Main title for plot (optional), default is "Jurors' Verdict Preferences, P(g)".}

\item{xlab}{X-axis label for plot (optional), default is "Jury Verdict Probabilities, P(G)".}

\item{ylab}{Main title for plot (optional), default is no main title.}
}
\value{
No return
}
\description{
Creates the shell of a plot showing relationship between jury pool preferences and jury verdict probabilities,
             optional argument to modify main, xlab, and ylab labels, includes grid lines.
}
\examples{
   library(sate)
   basic.plot.grid()

   basic.plot.grid(main="Death Sentencing Analysis", xlab="Jurors' Sentencing Preferences, P(d)",
                   ylab="Jury Verdict Probabilities, P(D)")
}
