% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI90.r
\name{CI90}
\alias{CI90}
\title{Calculates the 90 percent confidence interval of a proportion.}
\usage{
CI90(p, se)
}
\arguments{
\item{p}{The sample proportion (of jurors who favor a guilty verdict).}

\item{se}{The standard error of the sample proportion, p.}
}
\value{
Returns the 90 percent confidence interval as a list.
}
\description{
Calculates the 90 percent confidence interval of a proportion. 90 percent confidence interval used to test one-sided hypothesis at .05 level.
}
\examples{
   CI90(p=.5, se=.04)

   CI90(p=10/12, se=.02)
}
