% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{compact_harm_plot}
\alias{compact_harm_plot}
\title{Creates the shell of a plot used to display estimate of harm relative to harm threshold}
\usage{
compact_harm_plot()
}
\value{
No return
}
\description{
Creates the shell of a plot used for compact display estimate of harm estimate
             relative to harm thresholds.
}
\examples{
   library(sate)
   compact_harm_plot()
}
