% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.prop.r
\name{se.prop}
\alias{se.prop}
\title{Calculates the standard error of proportion.}
\usage{
se.prop(p, n)
}
\arguments{
\item{p}{The proportion (of jurors who favor a guilty verdict).}

\item{n}{The size of the sample used to estimate p.}
}
\value{
Returns the standard error of a sample proportion.
}
\description{
Calculates the standard error of proportion.
}
\examples{
   se.prop(p=.50, n=500)

   se.prop(p=10/12, n=400)
}
