% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.jury.point.r
\name{as.jury.point}
\alias{as.jury.point}
\title{Calculates probability jury finds defendant guilty based on juror preferences}
\usage{
as.jury.point(pg)
}
\arguments{
\item{pg}{The proportion of jurors who favor a guilty verdict.}
}
\value{
Returns the probability jury finds defendant guilty.
}
\description{
Calculates probability jury finds defendant guilty based on juror preferences. Does not estimate uncertainty.
}
\examples{
   as.jury.point(pg=.50)

   as.jury.point(pg=10/12)
}
